pkg/api: split into oss/enterprise
--- sturdy ---
user_id: cc33c18b-27e5-470d-8589-f9391ca37590
view_id: 7cc18852-5245-47e1-a119-d35364ed3156
workspace_id: 94a27834-303b-48d5-bfbb-76600296b1b6
