--- pkg/api/module.go
@@ -1,9 +1,0 @@
-package api
-
-import (
-	"mash/pkg/di"
-)
-
-var Module = di.NewModule(
-	di.Provides(ProvideAPI),
-)
