+++ pkg/github/db/module.go
@@ -1,0 +1,10 @@
+package db
+
+import "mash/pkg/di"
+
+var Module = di.NewModule(
+	di.Provides(NewGitHubInstallationRepo),
+	di.Provides(NewGitHubPRRepo),
+	di.Provides(NewGitHubRepositoryRepo),
+	di.Provides(NewGitHubUserRepo),
+)
